backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
opencv-python-headless
numpy
scikit-learn
passlib[bcrypt]
python-jose[cryptography]
Pillow
bcrypt
